ADD warning about single condition in readme
Some people has been tricked by the omitted curly bracket style in
passing parameter to methods. It produces a single parameter to any_of,
and thus a single condition joined by AND rather than OR.

Added a bit of warning about it in README until a rails warning is
issued when passing a single parameter.